Move base setup out of requests dir
