+++ drivers/hmis/spec/requests/hmis/sessions_controller_spec.rb
@@ -1,6 +1,6 @@
 require 'rails_helper'
 require_relative 'login_and_permissions'
-require_relative 'hmis_base_setup'
+require_relative '../../support/hmis_base_setup'
 
 RSpec.describe Hmis::SessionsController, type: :request do
   let(:user) { create :user }
